I've implemented Thrift report source and Thrift reporter interfaces.  This allows instrumented programs to use Facebook's Thrift RPC layer to transmit reports to the backend.
I also implemented a Proxy that can run on a remote node and collect reports from that node and forward them to the backend.  The protocol that should be used for those transfers is the Thrift protocol.

I also moved us away from using BasicConfigurator.configure() and instead using a real log4j config file.

git-svn-id: https://xtraceprj.svn.sourceforge.net/svnroot/xtraceprj/trunk@39 63ced123-6a40-0410-8707-b735ea0ac295